fix(indexer): SMI-6481 -- close governance round-3 findings; verify scripts/ with real tsc
Round 3 on 4301f6cea: 0 blockers, 4 should-fix, 4 nits. All 8 fixed. Two were
re-creations of the defect class the previous commit claimed to close, and BOTH
were found by running `tsc` directly on the file -- something I had not been
doing, because `npm run typecheck` structurally cannot see `scripts/`.

That is the headline. `tsconfig.json` is `"files": []` + `packages/` references,
so every "typecheck clean" claim in the previous two commits was true only of
`packages/`. Verified directly this time, per file:

- SF2: the round-2 `Number.isFinite` change to validateRow's success-path spread
  introduced a REAL type regression. `Number.isFinite` is declared
  `(number: unknown) => boolean` -- not a type predicate -- so it does not
  narrow, and the emitted object stopped being assignable to `SimRowResult`
  (TS2322). Runtime was always correct (the guard 30 lines above already
  rejected non-finite values), which is exactly why it was invisible. Now
  `typeof x === 'number' && Number.isFinite(x)`: same runtime, narrowing
  restored, still visibly matching the guard.
- SF1: the round-2 F3 fix -- replacing `as never` with a real type annotation --
  was incomplete. `cleanScan` omitted four required `MergedEdgeScanResult`
  fields and still "passed" typecheck. An unenforced annotation in an unchecked
  directory hides a shape mismatch exactly as well as the cast it replaced.

Also fixed:
- SF3: F13 fixed only one of the two unmocked files. `recheck.sibling-rescan.test.ts`
  also calls `runRecheck` -> `buildGitHubHeaders` and carried the identical live
  api.github.com exposure. The round-2 comment claiming this file "was the
  outlier" was false; both are mocked now and the comment names both.
- SF4: the F6/F7 truncation and `safeToString` fallback shipped with zero tests.
  Five added, including the null-prototype case where `String()` itself throws.
- NIT1: the number/bigint short-circuit returned before truncation (`10n ** 300n`
  is 301 chars). Single exit point now.
- NIT2: `slice` split surrogate pairs, emitting a lone `\uD83D` into an
  operator-facing message. Slices by code point.
- NIT3: "String(v) had an implicit bound" was true only for plain objects.
- NIT4: replaced the `startsWith('row_results.')` prefix-match total with an
  explicit `suppressedCount`. Correct today, but it silently coupled the
  operator-facing total to no other error string ever taking that prefix.

Separately: checkpoint.ts had crept back to 499/500 lines -- one from the
blocking gate, after two rounds of trimming my own comments. Trimming was not
holding, so `assertValidCheckpointShape` moved to
`smi5879-simulate-full.checkpoint-shape.ts`. checkpoint.ts is now 313 lines and
owns file I/O; the new module owns whole-checkpoint shape; per-row field
validation stays in `.checkpoint-row-shape.ts`.

Verified per file with `npx tsc --noEmit --strict ... <file>` across all 25
changed files: no type errors introduced. Three files still report pre-existing
errors (`recheck.test.ts`, `recheck.sibling-rescan.test.ts`,
`smi5879-simulate-full.classification.test.ts`) -- all at lines this PR never
touched, confirmed by diff-hunk inspection, and all instances of the SMI-6486
backlog.

2139 indexer tests pass. Prettier clean, lint clean outside the pre-existing
`.claude/skills` fixtures (SMI-6484), audit:standards unchanged at 83 / 10
pre-existing / 0 failed.
